+++ Cargo.toml
@@ -14,19 +14,25 @@ keywords = ["webauthn", "authentication"]
 categories = ["authentication", "web-programming"]
 license = "MPL-2.0"
 
+[features]
+core = ["openssl", "rand"]
+default = ["core"]
+
 [dependencies]
-base64 = "0.13"
-log = "0.4"
-nom = "4"
 serde = "1.0"
+nom = "4"
+base64 = "0.13"
+serde_derive = "1.0"
+thiserror = "1.0"
+
+log = "0.4"
 serde_bytes = "0.11"
 serde_cbor = "0.11"
-serde_derive = "1.0"
 serde_json = "1.0"
 # We could consider replacing this with openssl rand.
-openssl = "0.10"
-rand = "0.8"
-thiserror = "1.0"
+
+openssl = { version = "0.10", optional = true }
+rand = { version = "0.8", optional = true }
 
 [dev-dependencies]
 askama = "0.10"
