+++ src/error.rs
@@ -1,6 +1,7 @@
 //! Possible errors that may occur during Webauthn Operation processing
 
 use base64::DecodeError as b64DecodeError;
+#[cfg(feature = "core")]
 use openssl::error::ErrorStack as OpenSSLErrorStack;
 use serde_cbor::error::Error as CBORError;
 use serde_json::error::Error as JSONError;
@@ -135,6 +136,7 @@ pub enum WebauthnError {
     #[error("In parsing the attestation object, there was insufficient data")]
     ParseInsufficentBytesAvailable,
 
+    #[cfg(feature = "core")]
     #[error("An openSSL Error has occured")]
     OpenSSLError(#[from] OpenSSLErrorStack),
